Add(variables): Working local variables management (without set/unset as tcsh), env variables access and the majority of scripting variables
