+++ include/mysh.h
@@ -86,4 +86,12 @@ int history(char *line, int *error);
 char **get_history_array(void);
 int right_redirection(char *file_path, special_type_e type);
 
+//variables
+char *find_local_variable(char *str, int i);
+char *get_var(char *str, int i);
+char *get_var_name(char *str, int i);
+char *find_num_variable(char *str, char **wa, int i, char **env);
+char *find_special_variable(char *str, int i, char **env, int *error);
+bool detect_variable_attribution(char **line);
+
 #endif
